requirements.txt
streamlit>=1.36
openai>=1.30
python-dotenv>=1.0
pandas>=2.2
pillow>=10.3
supabase>=2.6
requests>=2.31
